+++ src/js/imagery/mapbox-attribution.js
