+++ src/js/password-request.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import {NavigationBar} from "./components/PageComponents";
-import {getQueryString} from "./utils/textUtils";
+import {getQueryString} from "./utils/generalUtils";
 
 class PasswordRequest extends React.Component {
     constructor(props) {
