--- src/js/utils/fileUtils.js
@@ -1,5 +1,0 @@
-export function encodeFileAsBase64(file, callback) {
-    const reader = new FileReader();
-    reader.onloadend = () => callback(reader.result);
-    reader.readAsDataURL(file);
-}
