+++ src/js/utils/mercator.js
@@ -28,8 +28,8 @@ import {BingMaps, Cluster, TileWMS, Vector as VectorSource, XYZ} from "ol/source
 import {Circle as CircleStyle, Icon, Fill, Stroke, Style, Text as StyleText, RegularShape} from "ol/style";
 import {fromLonLat, transform, transformExtent} from "ol/proj";
 import {fromExtent, fromCircle} from "ol/geom/Polygon";
-import {formatDateISO} from "./dateUtils";
-import {mapboxAttributionText} from "./mapbox-attribution";
+import {formatDateISO} from "./generalUtils";
+import {mapboxAttributionText} from "../imagery/mapbox-attribution";
 
 /******************************************************************************
 ***
