--- src/js/utils/projectUtils.js
@@ -1,3 +1,0 @@
-export const plotLimit = 5000;
-export const perPlotLimit = 200;
-export const sampleLimit = 50000;
